Add margin calculation and shared exchange-rate subscription facade
